requirements.txt
# Requirements for Non-Stationary MDP Research
# Install with: pip install -r requirements.txt

# Core RL
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Logging
wandb>=0.15.0
tensorboard>=2.14.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Configuration
pyyaml>=6.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Video Recording (optional)
moviepy>=1.0.3
imageio>=2.31.0
